Removed client.py
